@@ -42,7 +42,7 @@ public class JsonParser {
     try {
       jObj = new JSONObject(json);
     } catch (JSONException e) {
-      Log.e("JSON Parser", "Error parsing data " + e.toString());
+      Log.e("JSON Parser", "Error parsing data " + e);
     }
 
     // return JSON String
